move benchmark subcommand into cache subcommand
